NEW: Light radius + shadow sample added to parsing

+++ includes/scene_structs.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   scene_structs.h                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: madelvin <madelvin@student.42.fr>          +#+  +:+       +#+        */
+/*   By: lilefebv <lilefebv@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/01 17:03:30 by lilefebv          #+#    #+#             */
-/*   Updated: 2025/07/26 17:09:46 by madelvin         ###   ########.fr       */
+/*   Updated: 2025/07/28 16:44:21 by lilefebv         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -139,7 +139,7 @@ typedef struct s_light
 	t_vec3	position;
 	double	brightness;
 	t_color	color;
-	size_t	shadow_sample;
+	int		shadow_sample;
 	float	radius;
 	double	shadow_factor;
 }	t_light;
